lineBot_comedySearchShow/requirements.txt
line-bot-sdk
flask
gunicorn
gspread
google-auth